+++ src/data/corpus_sqlite.h
@@ -77,7 +77,7 @@ public:
         maxDims.resize(ex.size(), 0);
       for(size_t i = 0; i < ex.size(); ++i) {
         if(ex[i].size() > (size_t)maxDims[i])
-          maxDims[i] = ex[i].size();
+          maxDims[i] = (int)ex[i].size();
       }
       sentenceIds.push_back(ex.getId());
     }
